Nest SearchProcessor classes in their Factories for a more compact code
