+++ buffer/src/main/java/io/netty/buffer/search/KmpSearchProcessorFactory.java
@@ -14,16 +14,55 @@
  */
 package io.netty.buffer.search;
 
+import io.netty.util.internal.PlatformDependent;
+
 /**
- * Factory that creates {@link KmpSearchProcessor}.
+ * Implements
+ * <a href="https://en.wikipedia.org/wiki/Knuth%E2%80%93Morris%E2%80%93Pratt_algorithm">Knuth-Morris-Pratt</a>
+ * string search algorithm.
  * Use static {@link AbstractSearchProcessorFactory#newKmpSearchProcessorFactory}
  * to create an instance of this factory.
- * @see SearchProcessorFactory
+ * Use {@link KmpSearchProcessorFactory#newSearchProcessor} to get an instance of {@link io.netty.util.ByteProcessor}
+ * implementation for performing the actual search.
+ * @see AbstractSearchProcessorFactory
  */
 public class KmpSearchProcessorFactory extends AbstractSearchProcessorFactory {
 
     private final int[] jumpTable;
     private final byte[] needle;
+
+    public static class Processor implements SearchProcessor {
+
+        private final byte[] needle;
+        private final int[] jumpTable;
+        private long currentPosition;
+
+        Processor(byte[] needle, int[] jumpTable) {
+            this.needle = needle;
+            this.jumpTable = jumpTable;
+        }
+
+        @Override
+        public boolean process(byte value) {
+            while (currentPosition > 0 && PlatformDependent.getByte(needle, currentPosition) != value) {
+                currentPosition = PlatformDependent.getInt(jumpTable, currentPosition);
+            }
+            if (PlatformDependent.getByte(needle, currentPosition) == value) {
+                currentPosition++;
+            }
+            if (currentPosition == needle.length) {
+                currentPosition = PlatformDependent.getInt(jumpTable, currentPosition);
+                return false;
+            }
+
+            return true;
+        }
+
+        @Override
+        public void reset() {
+            currentPosition = 0;
+        }
+    }
 
     KmpSearchProcessorFactory(byte[] needle) {
         this.needle = needle.clone();
@@ -42,11 +81,11 @@ public class KmpSearchProcessorFactory extends AbstractSearchProcessorFactory {
     }
 
     /**
-     * Returns a new {@link KmpSearchProcessor}.
+     * Returns a new {@link Processor}.
      */
     @Override
-    public KmpSearchProcessor newSearchProcessor() {
-        return new KmpSearchProcessor(needle, jumpTable);
+    public Processor newSearchProcessor() {
+        return new Processor(needle, jumpTable);
     }
 
 }
